requirements.txt
numpy
scipy
scikit-learn
pythran
pytorch
skorch
cvxpy
pyro
networkx